chore: DTO 정리
- DTO 관련 도메인 레포지토리로 묶어서 정리
- 사용되지 않는 DTO 삭제
