--- src/main/java/com/healthier/diagnosis/dto/QuestionRequestDto.java
@@ -1,24 +1,0 @@
-package com.healthier.diagnosis.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.databind.PropertyNamingStrategies;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.validation.constraints.NotNull;
-
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class QuestionRequestDto {
-    @NotNull
-    private String questionId;
-
-    @NotNull
-    private int answerId;
-}
