+++ src/main/java/com/healthier/diagnosis/dto/headache/HeadacheResponse.java
@@ -1,18 +1,18 @@
-package com.healthier.diagnosis.dto.headache.commonQuestion;
+package com.healthier.diagnosis.dto.headache;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import lombok.*;
+import lombok.Builder;
+import lombok.Data;
 
 import java.util.List;
 
-@Getter
+@Data
 @Builder
-@AllArgsConstructor
-@NoArgsConstructor
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class PrimaryHeadacheRequest {
-    private List<QnARequest> questions;
-    private int isChronic;
+public class HeadacheResponse {
+    private int type;
     private String message;
+    private List<QuestionDto> questions;
+    private ResultDto result;
 }
