--- src/main/java/com/healthier/diagnosis/dto/headache/commonQuestion/HeadacheResponse.java
@@ -1,20 +1,0 @@
-package com.healthier.diagnosis.dto.headache.commonQuestion;
-
-import com.fasterxml.jackson.databind.PropertyNamingStrategies;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import com.healthier.diagnosis.dto.headache.QuestionDto;
-import com.healthier.diagnosis.dto.headache.ResultDto;
-import lombok.Builder;
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-@Builder
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class HeadacheResponse {
-    private int type;
-    private String message;
-    private List<QuestionDto> questions;
-    private ResultDto result;
-}
