--- src/main/java/com/healthier/diagnosis/dto/headache/commonQuestion/QnARequest.java
@@ -1,18 +1,0 @@
-package com.healthier.diagnosis.dto.headache.commonQuestion;
-
-import com.fasterxml.jackson.databind.PropertyNamingStrategies;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class QnARequest {
-    private int questionId;
-    private int answerId;
-}
