+++ src/main/java/com/healthier/diagnosis/dto/headache/primaryHeadacheNext/PrimaryHeadacheNextRequest.java
@@ -1,4 +1,4 @@
-package com.healthier.diagnosis.dto.headache.commonQuestion;
+package com.healthier.diagnosis.dto.headache.primaryHeadacheNext;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
