+++ src/main/java/com/healthier/diagnosis/dto/headache/redFlagSign/RedFlagSignResponse.java
@@ -1,4 +1,4 @@
-package com.healthier.diagnosis.dto.headache.commonQuestion;
+package com.healthier.diagnosis.dto.headache.redFlagSign;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
