+++ src/main/java/com/healthier/diagnosis/dto/question/QuestionDto.java
@@ -1,10 +1,7 @@
-package com.healthier.diagnosis.dto;
+package com.healthier.diagnosis.dto.question;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import com.healthier.diagnosis.domain.question.Answer;
-import lombok.Builder;
-import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
