+++ src/main/java/com/healthier/diagnosis/dto/question/saveDiagnosis/SaveDiagnosisResponseDto.java
@@ -1,4 +1,4 @@
-package com.healthier.diagnosis.dto;
+package com.healthier.diagnosis.dto.question.saveDiagnosis;
 
 import com.healthier.diagnosis.domain.user.User;
 import lombok.*;
