+++ src/main/java/com/healthier/diagnosis/service/QuestionService.java
@@ -4,7 +4,9 @@ import com.healthier.diagnosis.domain.question.Answer;
 import com.healthier.diagnosis.domain.question.Question;
 import com.healthier.diagnosis.domain.log.Log;
 import com.healthier.diagnosis.domain.user.Track;
-import com.healthier.diagnosis.dto.*;
+import com.healthier.diagnosis.dto.question.*;
+import com.healthier.diagnosis.dto.question.decisiveQuestion.DecisiveQuestionRequestDto;
+import com.healthier.diagnosis.dto.question.headache.HeadacheDefaultQuestionAfterRequestDto;
 import com.healthier.diagnosis.exception.CustomException;
 import com.healthier.diagnosis.exception.ErrorCode;
 import com.healthier.diagnosis.repository.DiagnosisLogRepository;
