+++ src/main/java/com/healthier/diagnosis/service/UserService.java
@@ -4,8 +4,8 @@ import com.healthier.diagnosis.domain.diagnosis.Diagnosis;
 import com.healthier.diagnosis.domain.user.User;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.healthier.diagnosis.domain.oauth.KakaoProfile;
-import com.healthier.diagnosis.dto.SaveDiagnosisRequestDto;
-import com.healthier.diagnosis.dto.SaveDiagnosisResponseDto;
+import com.healthier.diagnosis.dto.question.saveDiagnosis.SaveDiagnosisRequestDto;
+import com.healthier.diagnosis.dto.question.saveDiagnosis.SaveDiagnosisResponseDto;
 import com.healthier.diagnosis.exception.CustomException;
 import com.healthier.diagnosis.exception.ErrorCode;
 import com.healthier.diagnosis.repository.DiagnosisRepository;
